+++ package.json
@@ -9,7 +9,7 @@
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
-    "@types/googlemaps": "^3.43.3",
+    "@types/google-maps": "^3.2.3",
     "@types/humps": "^2.0.1",
     "@types/jest": "^27.4.0",
     "@types/react": "^17.0.39",
