+++ src/pages/landing/Login/Login.tsx
@@ -1,19 +1,17 @@
 import { Navigate } from 'react-router';
-import { MenuItem, FormStatus } from 'components/common';
+import { FormStatus } from 'components/common';
+import Header from 'components/Layout/Header';
 import { Mobile } from 'components/Layout/Mobile';
 import { HappySmile } from 'components/Layout/HappySmile';
-import { useDispatch, useSession, useTranslation } from 'hooks';
-import { login } from 'state/actions/user-actions';
 import { LoginForm } from './components/LoginForm';
 import routesPaths from 'constants/routes-paths-constant';
 import testIds from 'constants/test-ids-constant';
 import { loginI18n } from 'constants/i18n-constant';
+import useLogin from './useLogin';
 import './login.scss';
 
 const Login = () => {
-  const handleSubmit = useDispatch(login);
-  const { authenticated } = useSession();
-  const t = useTranslation();
+  const { handleSubmit, authenticated, t } = useLogin();
 
   if (authenticated) {
     return <Navigate to={routesPaths.index} />;
@@ -22,11 +20,7 @@ const Login = () => {
   return (
     <article className="two-column-layout-wrap" data-testid={testIds.LOGIN_PAGE}>
       <div className="left">
-        <div className="header">
-          <div className="header__item">
-            <MenuItem />
-          </div>
-        </div>
+        <Header />
         <HappySmile />
         <FormStatus />
         <h1 className="letter-spacing">{t(loginI18n.PAGE_TITLE)}</h1>
